pyproject.toml
[project]
name = "anime-stock"
version = "0.1.0"
description = "Animetrics AI - Anime stock market intelligence with sentiment analysis and ML predictions"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Oleksandr Tretiak" }]
keywords = ["anime", "stock", "sentiment", "ml", "finance"]

dependencies = [
    "yfinance>=0.2.36",
    "feedparser>=6.0.10",
    "pandas>=2.2.0",
    "openai>=1.54.0",
    "streamlit>=1.40.0",
    "plotly>=5.24.0",
    "scikit-learn>=1.4.0",
    "mysql-connector-python>=8.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
]

[project.scripts]
anime-stock = "anime_stock.main:main"
anime-collect = "anime_stock.scripts.daily_collect:main"
anime-dashboard = "anime_stock.dashboard.app:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/anime_stock"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=anime_stock"
